# apidoc/swagger.yaml
openapi: 3.0.0
info:
  title: Lms Api's
  version: 1.0.0
  description: "Node.js application with typescript,javascript"
servers:
  - url: "http://localhost:3000/api/v1"
    description: SwaggerHub API Auto Mocking-LOCAL
...
paths:
  /users/google-sign:
    post:
      tags:
        - "Users"
      summary: "signup with google"
      description: ""
      operationId: "google_signup"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: "string"
                  example: "uejicqysgxb93ns.mbjksx"
      responses:
        200:
          description: "successful operation"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "user not found"
...
components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: "object"
      required:
        - "name"
        - "email"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
          example: "Vivan"
        email:
          type: "string"
          example: "vivan@gmail.com"
        google-signup:
          type: "boolean"
          example: "false"
        google-Id:
          type: "string"
          example: "IHJJH7JJJ"
        role:
          type: "string"
          example: "employee"
        active:
          type: boolean
          example: true
        token:
          type: "string"
          example: null
        reportee:
          type: "string"
          example: "5shdgfa6asjdfv6sasdfgsid"
...